requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
python-telegram-bot==13.6
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==2.1
redis==3.5.3
youtube_dl